requirements.txt
flask==2.3.3
flask-restful==0.3.10
python-dotenv==1.0.0
pika==1.3.2
requests==2.31.0
Pillow==10.1.0